Create BeautifulSnowFlakes.cpp
A connected component of a graph is called a snowflake if it contains exactly one node that has a degree not equal to 1.
A snowflake is Beautiful if there are no other snowflakes in the graph that are the same as it. Two snowflakes are the same if they have the same number of nodes.
Given a graph, how many Beautiful snowflakes does it contain?