Create MinimumJumps.cpp
You are given N buildings numbered from 1 to N and M pairs of building denoting which pair of buildings are close enough to jump from one another. Given current building you are at and final building where you want to be at, find out minimum number of jumps required. If it is not possible to reach to final building, print 0.