Create MeltingIcecream_II.cpp
You and your brother decided to buy ice cream to relax on a summer evening. After buying, you realise that the ice cream is melting and you must reach your home before `k` minutes to save the ice cream.

You are given a grid that represents your locality. In your locality, there are roads that are represented by ∗ and buildings that are represented by `#`. The coordinates of the ice cream shop are given by i, and your home is given by h.

In 1 minute, you can move in any of the 4 directions UP, DOWN, LEFT, and RIGHT to the new cell provided it lies within the grid, and the new cell is a road.