Create ProgrammingLibraries.cpp
You are writing a huge program and are using lots of programming libraries in it. Some libraries depend upon other libraries to run. A library will not install until all the libraries that it depends upon are installed. You have to find out the order in which to install the libraries so that all the libraries get installed without any error.

It is guaranteed that the problem can be solved for the given input.